[v2.0.0 alpha]: separate observer and subscriber thread
In pervious testings, we noticed that the keyboard hook lost
functionality in some case. The issue may happend because the
do the event handler in the same thread with the keyboard event
monitor. The keyboard event didn't pass into next hook timely.
Then the windows kernel remove the abnormal keyboard hook to
ensure the system working. Therefore, we separate the event handling
and event monitor.